refactor (LeagueUtil) split off MatchUtil and EndpointUtil * move EndpointUtil into /util
