+++ lib/Endpoint.js
@@ -1,7 +1,7 @@
 const ApiRequest = require('./ApiRequest');
 
 const ParameterError = require('./errors/ParameterError');
-const EndpointUtil = require('./EndpointUtil');
+const EndpointUtil = require('./util/EndpointUtil');
 const RegionAndPlatformUtil = require('./util/RegionAndPlatformUtil');
 
 /**
