+++ lib/LeagueUtil.js
@@ -1,6 +1,4 @@
 const deepmerge = require('deepmerge');
-const fs = require('fs');
-const path = require('path');
 
 const RegionAndPlatformUtil = require('./util/RegionAndPlatformUtil');
 
@@ -9,40 +7,8 @@ const RateLimiter = require('./RateLimiter');
 class LeagueUtil {
 	// TODO: maybe split into more specific classes instead of "Util" class where everything gets dumped into
 
-
-	/**
-	 *
-	 * @param {string} gameVersion Versioning string from MatchDto.gameVersion
-	 * @param {string[]} versions Array with valid ddragon version {@link StaticDataEndpoint.gettingVersions}
-	 * @return {string} The data dragon version correlating to the gameVersion
-	 */
-	static getVersionForGameVersion(gameVersion, versions) {
-		let indexFirstDot = gameVersion.indexOf('.');
-		let indexSecondDot = gameVersion.substr(indexFirstDot + 1).indexOf('.') + indexFirstDot + 1;
-		let gameVersionMajorMinor = gameVersion.substr(0, indexSecondDot);
-
-		const ddVersion = versions.find(version => {
-			return version.indexOf(gameVersionMajorMinor) === 0;
-		});
-
-		if (ddVersion === null) {
-			throw new Error(`LeagueUtil.getDdragonVersionForMatchVersion(): no ddragon Version found from gameVersion: ${gameVersionMajorMinor}`);
-		} else {
-			console.log(`LeagueUtil.getDdragonVersionForMatchVersion() version is resolved: ${ddVersion} gameVersion: ${gameVersion} majorMinor: ${gameVersionMajorMinor}`);
-			return ddVersion;
-		}
-	}
-
 	static getConfig(options = {}) {
 		return deepmerge(require('./config'), options);
-	}
-
-	/** Returns names of all implemented Endpoints */
-	static getEndpointNames() {
-		let endpointPath = path.resolve(__dirname + '/endpoints');
-		console.log(endpointPath);
-		let filenames = fs.readdirSync(endpointPath);
-		return filenames.map(filename => filename.replace('Endpoint.js', ''));
 	}
 
 	/**
