+++ lib/config.js
@@ -1,5 +1,5 @@
 const RequestCache = require('./RequestCache');
-const LeagueUtil = require('./LeagueUtil');
+const EndpointUtil = require('./util/EndpointUtil');
 /**
  * @type {LeagueJSDefaultConfig}
  */
@@ -44,7 +44,7 @@ let LeagueJSDefaultConfig = {
 		'numMaxRetries': 3,
 		'intervalRetryMS': 500,
 		// per default, retry is enabled on all endpoints
-		'retryEndpoints': LeagueUtil.getEndpointNames()
+		'retryEndpoints': EndpointUtil.getEndpointNames()
 	}
 };
