+++ lib/endpoints/ChampionMasteryEndpoint.js
@@ -1,6 +1,5 @@
 const Endpoint = require ('../Endpoint');
-const EndpointUtil = require('../EndpointUtil');
-const LeagueUtil = require('../LeagueUtil');
+const EndpointUtil = require('../util/EndpointUtil');
 
 /**
  * Endpoint to receive ChampionMastery information
