+++ lib/endpoints/LeagueEndpoint.js
@@ -1,6 +1,5 @@
 const Endpoint = require ('../Endpoint');
-const EndpointUtil = require('../EndpointUtil');
-const LeagueUtil = require('../LeagueUtil');
+const EndpointUtil = require('../util/EndpointUtil');
 
 /***
  * Endpoint to receive ranked information about leagues by queue or for specific summoners
