+++ lib/endpoints/SpectatorEndpoint.js
@@ -1,5 +1,5 @@
 const Endpoint = require ('../Endpoint');
-const EndpointUtil = require('../EndpointUtil');
+const EndpointUtil = require('../util/EndpointUtil');
 
 /***
  * Endpoint to receive live match information
