+++ lib/endpoints/StaticDataEndpoint.js
@@ -1,5 +1,5 @@
 const Endpoint = require('../Endpoint');
-const EndpointUtil = require('../EndpointUtil');
+const EndpointUtil = require('../util/EndpointUtil');
 const ParameterUtil = require('../util/ParameterUtil');
 
 // const ParameterError = require('../errors/ParameterError');
@@ -28,7 +28,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 *
 	 * @param {string | string[]} options.tags Tags to return additional data.
 	 *
@@ -88,7 +88,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p
 	 *
 	 * @param {string | string[]} options.tags Tags to return additional data.
 	 *
@@ -142,7 +142,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 *
 	 * @param {string | string[]} options.tags Tags to return additional data.
 	 *
@@ -202,7 +202,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 *
 	 * @param {string | string[]} options.tags Tags to return additional data.
 	 *
@@ -262,7 +262,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 * @param {string} options.locale
 	 * <p>Locale code for returned data (e.g., en_US, es_ES).
 	 * If not specified, the default locale for the region is used.</p>
@@ -300,7 +300,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 * @param {string} options.locale
 	 * <p>Locale code for returned data (e.g., en_US, es_ES).
 	 * If not specified, the default locale for the region is used.</p>
@@ -328,7 +328,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 *
 	 * @param {string | string[]} options.tags Tags to return additional data.
 	 *
@@ -374,7 +374,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 *
 	 * @param {string | string[]} options.tags Tags to return additional data.
 	 *
@@ -415,7 +415,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 * @param {string} options.locale
 	 * <p>Locale code for returned data (e.g., en_US, es_ES).
 	 * If not specified, the default locale for the region is used.</p>
@@ -455,7 +455,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 *
 	 * @param {string | string[]} options.tags Tags to return additional data.
 	 *
@@ -498,7 +498,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 *
 	 * @param {string | string[]} options.tags Tags to return additional data.
 	 *
@@ -545,7 +545,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 *
 	 * @param {string | string[]} options.tags Tags to return additional data.
 	 *
@@ -603,7 +603,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * <p>If not specified, the latest version for the region is used.</p>
 	 * <p>List of valid versions can be obtained from the {@link StaticDataEndpoint.gettingVersions} endpoint.</p>
 	 * <p>Data dragon version for a specific {@link MatchDto.gameVersion} received from MatchEndpoint
-	 * can be obtained from {@link LeagueUtil#getVersionForGameVersion}</p>
+	 * can be obtained from {@link MatchUtil#getVersionForGameVersion}</p>
 	 *
 	 * @param {string | string[]} options.tags Tags to return additional data.
 	 *
