+++ lib/util/EndpointUtil.js
@@ -1,9 +1,20 @@
-const ParameterError = require('./errors/ParameterError');
-const SummonerUtil = require('./util/SummonerUtil');
-const RateLimiter = require('./RateLimiter');
-const RegionAndPlatformUtil = require('./util/RegionAndPlatformUtil');
+const fs = require('fs');
+const path = require('path');
+
+const ParameterError = require('../errors/ParameterError');
+const SummonerUtil = require('./SummonerUtil');
+const RateLimiter = require('../RateLimiter');
+const RegionAndPlatformUtil = require('./RegionAndPlatformUtil');
 
 class EndpointUtil {
+
+	/** Returns names of all implemented Endpoints */
+	static getEndpointNames() {
+		let endpointPath = path.resolve(__dirname, '../endpoints');
+		console.log(endpointPath);
+		let filenames = fs.readdirSync(endpointPath);
+		return filenames.map(filename => filename.replace('Endpoint.js', ''));
+	}
 
 	static buildQueryStringFromOptions(options) {
 		let paramsArray = Object.keys(options).map(optionKey => {
